REALIZATION - plus indentation
I realized that the ostensibly "old" clock tower check code is not old at all, but is actually meant to be directly appended to the end of the previous code being nested in that if statement. Therefore, the "old" code, is not old at all, but is instead new code that continues to elaborate on what is already being done. Sweetsauce. Also, I did add indentations for more clarity, since I did not indent my $elseif or $else statements before writing this code.