Continuing Writing Relapse Code
Continuing writing this relapse checker code. It's amazing to think that all of this work that I'm doing is just to record an accurate date. It doesn't even involve making calculations about the date and ensuring that they're accurate based on amount of time that has passed. I feel like I have a lot of coding to do, but I don't mind that there is a lot. I care most about whether or not it will work. I'm confident in my coding skills that it will work.