Update newcore
I think next I need to update it to include leap year days, following which I need to create previous day without min changes that are on the 1st of the month, followed by next day without min changes that are at the end of the month, followed by previous day with min changes that are not on the first of the month, followed by ones that are on the first of the month, as well as next day with min changes that are not at the end of the month, followed by ones that are on the end of the month. 

That should do it.